auto register data model
